created new component, refactoring app.jsx
